finish

+++ src/s21_decimal/s21_set_sign.c
@@ -2,9 +2,9 @@
 
 void s21_set_sign(s21_decimal *value, int sign) {
   if (sign == 0) {
-    value->bits[3] &= ~(1 << S21_DEC_SIGN_SHIFT);
+    value->bits[3] &= ~((uint32_t)1 << S21_DEC_SIGN_SHIFT);
   } else if (sign == 1) {
-    value->bits[3] |= (1 << S21_DEC_SIGN_SHIFT);
+    value->bits[3] |= ((uint32_t)1 << S21_DEC_SIGN_SHIFT);
   }
   return;
 }
